requirements.txt
snowflake-connector-python
pandas
openpyxl
python-dotenv
